Changed a magic mystery value from 24 to 16.
There was a logic bug in MOJOSHADER_parsePreshader() that was fixed in the
merge, but it caused this mystery value to be wrong (before, we would step
too far in an array and this value was covering for it).

We still don't know what's at the end of this array, but at least we aren't
crashing now and it appears to work as well as before. It's still a mystery.